refactor: move trusted peers resolver to peers.rs
