+++ crates/net/peers/Cargo.toml
@@ -25,8 +25,6 @@ serde_with.workspace = true
 thiserror.workspace = true
 url.workspace = true
 tokio = { workspace = true, optional = true }
-futures = { workspace = true, optional = true }
-tracing.workspace = true
 
 [dev-dependencies]
 alloy-primitives = { workspace = true, features = ["rand"] }
@@ -48,4 +46,4 @@ std = [
     "serde_json/std",
 ]
 secp256k1 = ["dep:secp256k1", "enr/secp256k1"]
-net = ["std", "dep:tokio", "tokio?/net", "tokio/time", "dep:futures"]
+net = ["std", "dep:tokio", "tokio?/net"]
