+++ crates/net/peers/src/lib.rs
@@ -74,7 +74,7 @@ pub mod node_record;
 pub use node_record::{NodeRecord, NodeRecordParseError};
 
 pub mod trusted_peer;
-pub use trusted_peer::{TrustedPeer, TrustedPeersResolver};
+pub use trusted_peer::TrustedPeer;
 
 mod bootnodes;
 pub use bootnodes::*;
